refactor: ResultDealer 삭제 및 CardHand로 이동
